refactor voucher and implement authorization

+++ voucher-service/src/main/java/com/mahoni/voucherservice/voucher/service/VoucherService.java
@@ -1,12 +1,16 @@
 package com.mahoni.voucherservice.voucher.service;
 
+import com.mahoni.voucherservice.merchant.model.Merchant;
+import com.mahoni.voucherservice.merchant.service.MerchantService;
 import com.mahoni.voucherservice.voucher.dto.VoucherRequest;
-import com.mahoni.voucherservice.voucher.exception.VoucherAlreadyExistException;
 import com.mahoni.voucherservice.voucher.exception.VoucherNotFoundException;
 import com.mahoni.voucherservice.voucher.model.Voucher;
 import com.mahoni.voucherservice.voucher.repository.VoucherRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,20 +20,27 @@ import java.util.UUID;
 @Service
 @Slf4j
 public class VoucherService {
+
   @Autowired
   VoucherRepository voucherRepository;
 
+  @Autowired
+  MerchantService merchantService;
+
   public Voucher create(VoucherRequest voucher) {
-//    TODO: request user, continue if merchant
-    if (voucherRepository.findByName(voucher.getName()).isPresent() &&
-      voucherRepository.findByCode(voucher.getCode()).isPresent()) {
-      throw new VoucherAlreadyExistException(voucher.getName(), voucher.getCode());
-    }
-    return voucherRepository.save(new Voucher(voucher.getName(),
+    Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+    Merchant merchant = merchantService.getByUsername(auth.getName());
+    Voucher newVoucher = new Voucher(
+      voucher.getName(),
       voucher.getDescription(),
-      voucher.getCode(),
+      voucher.getType(),
+      voucher.getPoint(),
       voucher.getStartAt(),
-      voucher.getExpiredAt()));
+      voucher.getExpiredAt(),
+      merchant);
+    newVoucher.setQuantity(0);
+
+    return voucherRepository.save(newVoucher);
   }
 
   public Voucher getById(UUID id) {
@@ -49,6 +60,11 @@ public class VoucherService {
     if (voucher.isEmpty()) {
       throw new VoucherNotFoundException(id);
     }
+    Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+    if (!(voucher.get().getMerchant().getUsername().equals(auth.getName()) || auth.getAuthorities().stream()
+      .anyMatch(r -> r.getAuthority().equals("ADMIN")))) {
+      throw new AccessDeniedException("You don’t have permission to access this resource");
+    }
     voucherRepository.deleteById(id);
     return voucher.get();
   }
@@ -56,12 +72,17 @@ public class VoucherService {
   public Voucher update(UUID id, VoucherRequest newVoucher) {
     Optional<Voucher> voucher = voucherRepository.findById(id);
     if (voucher.isEmpty()) {
-      throw new VoucherNotFoundException();
+      throw new VoucherNotFoundException(id);
+    }
+    Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+    if (!(voucher.get().getMerchant().getUsername().equals(auth.getName()) || auth.getAuthorities().stream()
+      .anyMatch(r -> r.getAuthority().equals("ADMIN")))) {
+      throw new AccessDeniedException("You don’t have permission to access this resource");
     }
     Voucher updatedVoucher = voucher.get();
     updatedVoucher.setName(newVoucher.getName());
     updatedVoucher.setDescription(newVoucher.getDescription());
-    updatedVoucher.setCode(newVoucher.getCode());
+    updatedVoucher.setType(newVoucher.getType());
     updatedVoucher.setStartAt(newVoucher.getStartAt());
     updatedVoucher.setExpiredAt(newVoucher.getExpiredAt());
     return voucherRepository.save(updatedVoucher);
